Extract MetalKernelError to err file
