+++ candle-metal-kernels/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod err;
 pub mod kernels;
 pub mod metal;
 pub mod mlx_gemm;
@@ -5,6 +6,7 @@ pub mod sort;
 pub mod source;
 pub mod utils;
 
+pub use err::MetalKernelError;
 pub use kernels::Kernels;
 use metal::{
     BlitCommandEncoder, Buffer, CommandQueue, ComputeCommandEncoder, ComputePipeline, Device,
@@ -135,45 +137,6 @@ pub mod unary {
 }
 pub mod binary {
     ops!(add, sub, mul, div, min, max, eq, ne, le, lt, ge, gt);
-}
-
-#[derive(thiserror::Error, Debug)]
-pub enum MetalKernelError {
-    #[error("Could not lock kernel map: {0}")]
-    LockError(String),
-    #[error("Error while loading library: {0}")]
-    LoadLibraryError(String),
-    #[error("Error while loading function: {0}")]
-    LoadFunctionError(String),
-    #[error("Failed to create compute function")]
-    FailedToCreateComputeFunction,
-    #[error("Failed to create metal resource: {0}")]
-    FailedToCreateResource(String),
-    #[error("Failed to create pipeline")]
-    FailedToCreatePipeline(String),
-    #[error("Invalid matmul arguments {lhs_stride:?} {rhs_stride:?} {mnk:?}")]
-    MatMulNonContiguous {
-        lhs_stride: Vec<usize>,
-        rhs_stride: Vec<usize>,
-        mnk: (usize, usize, usize),
-    },
-    #[error("Sdpa {variation} head size was {got}, expectd {expected:?}")]
-    SdpaHeadSizeMismatch {
-        variation: &'static str,
-        got: usize,
-        expected: Vec<usize>,
-    },
-    #[error("Sdpa {variation} got dtype {got:?}")]
-    SdpaHeadDTypeMismatch {
-        variation: &'static str,
-        got: SdpaDType,
-    },
-}
-
-impl<T> From<std::sync::PoisonError<T>> for MetalKernelError {
-    fn from(e: std::sync::PoisonError<T>) -> Self {
-        Self::LockError(e.to_string())
-    }
 }
 
 #[allow(clippy::too_many_arguments)]
